LayoutTests/editing/caret/ios/fixed-caret-position-after-scroll-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS finalCaretRect.top - initialCaretRect.top is 500
